requirements.txt
numpy==1.24.3
opencv_python==4.7.0.72
python-dotenv==1.0.0
supervision==0.3.0
ultralytics==8.0.98
PyPDF2==3.0.1
